Error handling is now performed, by default, after one run loop iteration.
This is intended to function essentially as does -autorelease, by allowing the resolved
future to be passed out of its scope before it is actually resolved. This permits a method
to immediately return a future and resolve it as an error while giving the caller enough
time to register a failure handler with it before the resolution is triggered.